Pathway component
pathway component just with the handles
