+++ UOP1_Project/Assets/Prefabs/Pathway.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40e40b1ee07a76f46bd4bdb320066635
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
