--- UOP1_Project/Assets/Scripts/Editor/Pathway/PathWayNavMeshUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1114de2a61dc66e489f951c4ed04991f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
