--- UOP1_Project/Assets/Scripts/Editor/Pathway/PathwayNavMesh.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 89ec3dfe690498548a5787eb410bb855
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
