+++ UOP1_Project/Assets/Scripts/Pathway.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aaf94d94b26470e4cabe86e27f3fdb27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
